+++ extensions/obsidian/src/components/NoteList.tsx
@@ -77,9 +77,11 @@ export function NoteListItem(props: {
       }
       actions={
         <ActionPanel>
-          <OpenNoteActions note={note} vault={vault} actionCallback={actionCallback} />
-          <NoteActions note={note} vault={vault} actionCallback={actionCallback} />
-          {props.action && props.action(note)}
+          <React.Fragment>
+            <OpenNoteActions note={note} vault={vault} actionCallback={actionCallback} />
+            <NoteActions note={note} vault={vault} actionCallback={actionCallback} />
+            {props.action && props.action(note)}
+          </React.Fragment>
         </ActionPanel>
       }
     />